requirements.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
charset-normalizer==2.0.9
click==7.1.2
colorclass==2.2.0
colorlog==6.6.0
croniter==1.1.0
Deprecated==1.2.13
faust==1.10.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kafka-python==1.4.7
MarkupSafe==2.0.1
mode==4.3.2
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.21.4
opentracing==1.3.0
pandas==1.3.4
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
robinhood-aiokafka==1.1.6
six==1.16.0
SQLAlchemy==1.4.27
terminaltables==3.1.7
typing-extensions==4.0.1
venusian==1.2.0
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
